Abstract log logic to trait, set defaults in init
