+++ src/log/FileTarget.php
@@ -8,9 +8,7 @@
 namespace craft\log;
 
 use Craft;
-use craft\helpers\ArrayHelper;
-use yii\helpers\VarDumper;
-use yii\web\Request;
+use craft\base\LogTrait;
 
 /**
  * Class FileTarget
@@ -20,62 +18,24 @@ use yii\web\Request;
  */
 class FileTarget extends \yii\log\FileTarget
 {
-    /**
-     * @var bool Whether the user IP should be included in the default log prefix.
-     * @since 3.0.25
-     * @see prefix
-     */
-    public $includeUserIp = false;
+    use LogTrait;
 
     /**
      * @inheritdoc
      */
-    public function getMessagePrefix($message)
+    public function init()
     {
-        if ($this->prefix !== null) {
-            return call_user_func($this->prefix, $message);
-        }
+        parent::init();
 
-        if (Craft::$app === null) {
-            return '';
-        }
+        $generalConfig = Craft::$app->getConfig()->getGeneral();
+        $isConsoleRequest = Craft::$app->getRequest()->getIsConsoleRequest();
+        $this->fileMode = $generalConfig->defaultFileMode;
+        $this->dirMode = $generalConfig->defaultDirMode;
 
-        if ($this->includeUserIp) {
-            $request = Craft::$app->getRequest();
-            $ip = $request instanceof Request ? $request->getUserIP() : '-';
+        if ($isConsoleRequest) {
+            $this->logFile = '@storage/logs/console.log';
         } else {
-            $ip = '-';
+            $this->logFile = '@storage/logs/web.log';
         }
-
-        /* @var $user \yii\web\User */
-        $user = Craft::$app->has('user', true) ? Craft::$app->get('user') : null;
-        if ($user && ($identity = $user->getIdentity(false))) {
-            $userID = $identity->getId();
-        } else {
-            $userID = '-';
-        }
-
-        /* @var $session \yii\web\Session */
-        $session = Craft::$app->has('session', true) ? Craft::$app->get('session') : null;
-        $sessionID = $session && $session->getIsActive() ? $session->getId() : '-';
-
-        return "[$ip][$userID][$sessionID]";
-    }
-
-    /**
-     * @inheritdoc
-     */
-    protected function getContextMessage()
-    {
-        $context = ArrayHelper::filter($GLOBALS, $this->logVars);
-        $result = [];
-        $security = Craft::$app->getSecurity();
-
-        foreach ($context as $key => $value) {
-            $value = $security->redactIfSensitive($key, $value);
-            $result[] = "\${$key} = " . VarDumper::dumpAsString($value);
-        }
-
-        return implode("\n\n", $result);
     }
 }
